FEAT: implement logout menu component and refactor searchbar for improved user experience
